Code review part 2 - refactored cpu_usage_sampler into OS specific folders. Added stub/default implementation for Windows and MacOS
